Task #164639: Create edit and delete user UI.

See merge request research/web-educational-course-4-0-q1-22/web-educational-course-4-0-q1-22/web-educational-course-4-0-q1-22!7

+++ Lab2/Lab2/Services/UserService.cs
@@ -165,7 +165,7 @@ namespace Lab2.Services
                 throw new EntityNotFoundException();
             }
 
-            if (foundUser.ImageBlobKey != null)
+            if (foundUser.ImageBlobKey != null && file != null)
             {
                 string absoluteFilePath = $"{env.WebRootPath}/{foundUser.ImageBlobKey}";
                 File.Delete(absoluteFilePath);
